+++ .github/pull_request_template.md
@@ -1,31 +1,23 @@
 # Pull Request Template
 
-> `PR Template`은 유용한 PR을 통해 리뷰어와 유의미한 소통으로 더 많은 성장을 하기 위함을 목적으로 하고 있습니다. (삭제)
-> 아래의 템플릿은 그대로 사용하셔도 되고 선택적으로 참고만하셔도 됩니다. (삭제)
-> (삭제)가 표기된 가이드 문구는 PR 올리실때 모두 제거해주세요. (삭제)
-
 ## 1. 가이드
 
 ### 1-1. 배포
 
-> 배포된 페이지가 있다면 링크를 남겨주세요 (삭제)
 
-- [LMS](https://edu.nextstep.camp)
-- [CSS](next-step.github.io/shopping-cart-css)
+- [데모링크](https://react-shopping-cart-sooty-nine.vercel.app/)
+
 
 ### 1-2. 로컬 환경
 
-> 로컬 환경 확인이 필요하다면 예시와 같이 남겨주세요 (삭제)
 
 ```bash
 # install
 yarn
 
 # run
-yarn dev
+yarn start
 
-# test
-yarn test
 ```
 
 ## 2. 리뷰 요청 ✍️
@@ -48,9 +40,45 @@ yarn test
 
 ### 4-1. 필수 요구사항
 
-- [ ] `Storybook` 상호 작용 테스트
-- [ ] `MSW`를 활용한 `API mocking`
+- [x]  [MSW](https://mswjs.io/)를 활용한 **`API mocking`**
+- [x]  `Endpoint`만 변경하면 언제든 `Real API`를 바라볼 수 있다고 가정하고 상상합니다.
+    - [x]  `Real API` 없이 로컬에서만 동작하는 상태로 리뷰 받는 것이 기본 원칙입니다.
+
+
+- [x]  `로고`를 누르면 상품목록 페이지로 이동한다.
+- [x]  `장바구니` 버튼을 누르면 장바구니 페이지로 이동한다.
+- [x]  `주문목록` 버튼을 누르면 주문목록 페이지로 이동한다.
+
+- [x]  상품들은 n x 4 레이아웃으로 보여진다.
+- [x]  상품들에는 사진, 이름, 금액이 보여진다.
+- [x]  장바구니 버튼을 클릭하면 (**)
 
 ### 4-2. 선택 요구사항 (심화)
 
+### **상품상세**
+
+- [x] 페이지에는 상품 사진, 이름, 금액 정보가 보여진다.
+- [ ] 장바구니 버튼을 클릭하면 장바구니 페이지로 이동한다.
+- [ ] 장바구니 버튼을 클릭하면 해당 상품이 장바구니에 담긴다.
+
+### **주문 상세**
+
+- [ ] 주문 정보가 보여진다.
+- [ ] 장바구니 버튼을 클릭하면, 해당 상품이 장바구니에 담기고 장바구니 이동 선택 모달이 보여진다.
+    - [ ] 장바구니 이동 버튼을 누르면 장바구니 페이지로 이동한다.
+
+### **UX/UI**
+
 - [ ] 반응형 레이아웃을 구현한다.
+- [ ] 사용성 개선
+    - [ ] 사용자를 위한 로딩 환경 개선
+    - [ ] 상품이 없을 때와 같은 다양한 `Edge Case` 대응
+    - [ ] 반응형 레이아웃 구현
+    - [ ] 별도의 모바일 레이아웃 추가 제공
+- [ ] 매출 증대 및 마케팅을 위해 별도의 기능 구현 (별도의 API 없음)
+    - [ ] 브라우저 새로고침시 모든 상태 유지
+    - [ ] 흐름을 고려한 맞춤 큐레이팅 상품 추천 기능
+    - [ ] 구매 유도를 위한 상품 찜 페이지
+- [ ] 매출 증대 및 마케팅을 위한 별도의 도구 추가
+    - [ ] Google Analytics
+    - [ ] Google Tag Manager
