--- src/constant/router.ts
@@ -1,8 +1,0 @@
-export const ROUTE = {
-  HOME: '/',
-  DETAIL: '/products',
-  CART: '/cart',
-  ORDER: '/order',
-  ORDER_LIST: '/order-list',
-  ORDER_DETAIL: '/order-detail',
-};
