+++ src/index.tsx
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom/client';
 import App from './App';
 import './style/css/index.css';
 import { worker } from './mocks/browser';
-import { BrowserRouter } from 'react-router-dom';
+
 // if (process.env.NODE_ENV === 'development') {
 worker.start();
 // }
@@ -13,8 +13,6 @@ const root = ReactDOM.createRoot(
 );
 root.render(
   <React.StrictMode>
-    <BrowserRouter>
-      <App />
-    </BrowserRouter>
+    <App />
   </React.StrictMode>
 );
