+++ src/pages/Cart.tsx
@@ -1,12 +1,13 @@
 import SectionTitle from '../components/common/SectionTitle';
 import SectionCartList from '../components/cart/SectionCartList';
+import Layout from '../layout/Layout';
 
 const Cart = () => {
   return (
-    <>
+    <Layout>
       <SectionTitle text="장바구니" />
       <SectionCartList />
-    </>
+    </Layout>
   );
 };
