+++ src/pages/Home.tsx
@@ -1,11 +1,12 @@
 import React from 'react';
 import SectionProductList from '../components/home/SectionProductList';
+import Layout from '../layout/Layout';
 
 const Home = () => {
   return (
-    <>
+    <Layout>
       <SectionProductList />
-    </>
+    </Layout>
   );
 };
