+++ src/pages/ProductDetail.tsx
@@ -1,7 +1,8 @@
 import { useNavigate, useParams } from 'react-router-dom';
 import { useEffect, useState } from 'react';
 import { fetchProductDetail, ProductType } from '../api/product';
-import { ROUTE } from '../constant/router';
+import { ROUTE } from '../router';
+import Layout from '../layout/Layout';
 
 const ProductDetail = () => {
   const { id } = useParams();
@@ -25,31 +26,35 @@ const ProductDetail = () => {
   }, []);
 
   return (
-    <div className="product-detail-container">
-      <div className="flex-col-center w-520">
-        <img
-          className="w-480 h-480 mb-10"
-          src={productData?.imageUrl}
-          alt={productData?.name}
-        />
-        <div className="product-detail-info">
-          <span className="product-detail-info__name">{productData?.name}</span>
-          <hr className="divide-line-gray my-20" />
-          <div className="flex justify-between">
-            <span>금액</span>
-            <span className="product-detail-info__price">
-              {productData?.price.toLocaleString()}원
+    <Layout>
+      <div className="product-detail-container">
+        <div className="flex-col-center w-520">
+          <img
+            className="w-480 h-480 mb-10"
+            src={productData?.imageUrl}
+            alt={productData?.name}
+          />
+          <div className="product-detail-info">
+            <span className="product-detail-info__name">
+              {productData?.name}
             </span>
+            <hr className="divide-line-gray my-20" />
+            <div className="flex justify-between">
+              <span>금액</span>
+              <span className="product-detail-info__price">
+                {productData?.price.toLocaleString()}원
+              </span>
+            </div>
           </div>
+          <button
+            className="product-detail-button flex-center mt-20"
+            onClick={addToCart}
+          >
+            장바구니
+          </button>
         </div>
-        <button
-          className="product-detail-button flex-center mt-20"
-          onClick={addToCart}
-        >
-          장바구니
-        </button>
       </div>
-    </div>
+    </Layout>
   );
 };
